+++ MKButton.m
@@ -21,7 +21,7 @@
 #pragma mark Initialization
 - (id)init {
 	self = [super init];
-	if( self ) {
+	if (self) {
 		letter = nil;
 		keycode = nil;
 		xStart = 0;
@@ -32,9 +32,10 @@
 	return self;
 }
 
-- (id)initWithID:(int)aButtonID xStart:(int)aXStart xEnd:(int)aXEnd yStart:(int)aYStart yEnd:(int)aYEnd {
+- (id)initWithID:(NSInteger)aButtonID xStart:(NSInteger)aXStart xEnd:(NSInteger)aXEnd yStart:(NSInteger)aYStart
+	    yEnd:(NSInteger)aYEnd {
 	self = [self init];
-	if( self ) {
+	if (self) {
 		letter = nil;
 		keycode = nil;
 		buttonID = aButtonID;
@@ -56,9 +57,10 @@
 	return [[[[self class] alloc] init] autorelease];
 }
 
-+ (id)buttonWithID:(int)aButtonID xStart:(int)aXStart xEnd:(int)aXEnd yStart:(int)aYStart yEnd:(int)aYEnd {
++ (id)buttonWithID:(NSInteger)aButtonID xStart:(NSInteger)aXStart xEnd:(NSInteger)aXEnd yStart:(NSInteger)aYStart
+	      yEnd:(NSInteger)aYEnd {
 	return [[[[self class] alloc] initWithID:aButtonID xStart:aXStart xEnd:aXEnd yStart:aYStart yEnd:aYEnd]
-			autorelease];
+		autorelease];
 }
 
 + (id)buttonWithButton:(MKButton *)aButton letter:(NSString *)aLetter keycode:(NSString *)aKeycode {
@@ -76,8 +78,10 @@
 
 #pragma mark Utilities
 - (BOOL)containsPoint:(NSPoint)aPoint size:(NSSize)circleSize {
-	if( aPoint.x >= (xStart - (circleSize.width/2)) && (aPoint.x + (circleSize.width /2)) <= xEnd
-			&& aPoint.y >= (yStart - (circleSize.height/2)) && (aPoint.y+(circleSize.height/2)) <=yEnd ) {
+	if (aPoint.x >= (xStart - (circleSize.width/2))
+	    && (aPoint.x + (circleSize.width /2)) <= xEnd
+	    && aPoint.y >= (yStart - (circleSize.height/2))
+	    && (aPoint.y+(circleSize.height/2)) <=yEnd) {
 		return YES;
 	}
 	return NO;
@@ -86,7 +90,7 @@
 #pragma mark NSCopying
 - (id)copyWithZone:(NSZone *)zone {
 	return [[[self class] allocWithZone:zone] initWithID:[self buttonID] xStart:[self xStart] xEnd:[self xEnd]
-			yStart:[self yStart] yEnd:[self yEnd]];
+						      yStart:[self yStart] yEnd:[self yEnd]];
 }
 
 #pragma mark -
